fix sonarqube
